# ruler,receive,sidecar: StoreAPI Series encodes chunks to 120 samples instead of single, big one.