This is to have a unified chunk size and should reduce the load on queriers.
This also will be much more comparable when a chunk iterator will be done.

See following benchmark results for Receive (multiTSDB):

```
 benchstat -delta-test none _dev/bench_outs/0-receiveseries/benchBenchmarkMultiTSDBSeries.out _dev/bench_outs/1-receiveseries/benchBenchmarkMultiTSDBSeries.out
name                                                                                                  old time/op    new time/op    delta
MultiTSDBSeries/1000000SeriesWith1Samples/headOnly/4_TSDBs_with_1_samples,_250000_series_each-12         6.41s ± 0%     6.16s ± 0%   -3.85%
MultiTSDBSeries/1000000SeriesWith1Samples/blocksOnly/4_TSDBs_with_1_samples,_250000_series_each-12       5.77s ± 0%     6.16s ± 0%   +6.61%
MultiTSDBSeries/100000SeriesWith100Samples/headOnly/4_TSDBs_with_25_samples,_25000_series_each-12        3.68s ± 0%     3.96s ± 0%   +7.43%
MultiTSDBSeries/100000SeriesWith100Samples/blocksOnly/4_TSDBs_with_25_samples,_25000_series_each-12      4.04s ± 0%     4.02s ± 0%   -0.46%
MultiTSDBSeries/1SeriesWith10000000Samples/headOnly/4_TSDBs_with_2500000_samples,_1_series_each-12       1.53s ± 0%     1.57s ± 0%   +2.08%
MultiTSDBSeries/1SeriesWith10000000Samples/blocksOnly/4_TSDBs_with_2500000_samples,_1_series_each-12     1.67s ± 0%     1.66s ± 0%   -1.12%

name                                                                                                  old alloc/op   new alloc/op   delta
MultiTSDBSeries/1000000SeriesWith1Samples/headOnly/4_TSDBs_with_1_samples,_250000_series_each-12        4.08GB ± 0%    4.08GB ± 0%   +0.03%
MultiTSDBSeries/1000000SeriesWith1Samples/blocksOnly/4_TSDBs_with_1_samples,_250000_series_each-12      4.08GB ± 0%    4.08GB ± 0%   -0.01%
MultiTSDBSeries/100000SeriesWith100Samples/headOnly/4_TSDBs_with_25_samples,_25000_series_each-12       1.73GB ± 0%    1.72GB ± 0%   -0.37%
MultiTSDBSeries/100000SeriesWith100Samples/blocksOnly/4_TSDBs_with_25_samples,_25000_series_each-12     1.66GB ± 0%    1.67GB ± 0%   +0.57%
MultiTSDBSeries/1SeriesWith10000000Samples/headOnly/4_TSDBs_with_2500000_samples,_1_series_each-12      2.71GB ± 0%    2.47GB ± 0%   -8.68%
MultiTSDBSeries/1SeriesWith10000000Samples/blocksOnly/4_TSDBs_with_2500000_samples,_1_series_each-12    2.68GB ± 0%    2.46GB ± 0%   -8.14%

name                                                                                                  old allocs/op  new allocs/op  delta
MultiTSDBSeries/1000000SeriesWith1Samples/headOnly/4_TSDBs_with_1_samples,_250000_series_each-12         44.9M ± 0%     44.9M ± 0%   +0.00%
MultiTSDBSeries/1000000SeriesWith1Samples/blocksOnly/4_TSDBs_with_1_samples,_250000_series_each-12       44.9M ± 0%     44.9M ± 0%   -0.00%
MultiTSDBSeries/100000SeriesWith100Samples/headOnly/4_TSDBs_with_25_samples,_25000_series_each-12        27.1M ± 0%     27.1M ± 0%   -0.01%
MultiTSDBSeries/100000SeriesWith100Samples/blocksOnly/4_TSDBs_with_25_samples,_25000_series_each-12      27.1M ± 0%     27.1M ± 0%   +0.01%
MultiTSDBSeries/1SeriesWith10000000Samples/headOnly/4_TSDBs_with_2500000_samples,_1_series_each-12       1.02M ± 0%     1.69M ± 0%  +65.15%
MultiTSDBSeries/1SeriesWith10000000Samples/blocksOnly/4_TSDBs_with_2500000_samples,_1_series_each-12     1.02M ± 0%     1.69M ± 0%  +65.34%
```

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>

<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [x] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->