@bwplotka 2020-07-09
all is fine now, would be useful for future benchmarks. 

CC @brancz